+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2024-06-28
+
 ### Added
 
 - Added a helper function to get the logs from a given Pod
@@ -439,7 +441,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/giantswarm/clustertest/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/giantswarm/clustertest/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/giantswarm/clustertest/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/giantswarm/clustertest/compare/v1.5.0...v1.6.0